Remove 160 cap for RGB values on combine blend; many improvements to editor; improved selection accuracy; Removing some partually completed classes and functions from framework; Made a test map for a good snakey path...

+++ Apex/Classes/[[Editor]]/EditorMainMenu.cpp
@@ -283,13 +283,11 @@ void EditorMainMenu::Notify(void *pSender, const char *pNotification) {
             gWorldContainer->mGestureContainer->BaseLayout();
         }
 
-        if (pSender == mButtonAddPath) {
-            gEditor->AddPath();
-        }
+        if (pSender == mButtonAddPath) { gEditor->AddPath(); }
 
-        if (pSender == mButtonDeletePath) {
-            gEditor->DeleteCurrentPath();
-        }
+        if (pSender == mButtonDeletePath) { gEditor->DeleteCurrentPath(); }
+        if (pSender == mButtonExport) { gEditor->ExportImage(); }
+        if (pSender == mButtonSave) { gEditor->ExportMap(); }
 
     }
